Improve planner loop detection
The planner would skip any action that was already added to the plan,
however this would prevent repeat actions that modify the state
incrementally. The planner now skips an action if it has been used
before for the same state.

Change-type: patch
