+++ lib/planner/types.ts
@@ -1,5 +1,3 @@
-import { Action } from '../task';
-
 /**
  * Stats about the planning process
  */
@@ -19,40 +17,6 @@ export interface PlanningStats {
 	 */
 	time: number;
 }
-
-/**
- * Result of the planning process
- */
-export type PlanningResult<TState> =
-	| {
-			/**
-			 * Planning was successful
-			 */
-			success: true;
-			/**
-			 * The plan to get from the current state to the target state
-			 */
-			plan: Array<Action<TState>>;
-			/**
-			 * The expected state after applying the plan
-			 */
-			state: TState;
-			/**
-			 * Stats about the planning process
-			 */
-			stats: PlanningStats;
-	  }
-	| {
-			/**
-			 * Planning was not successful
-			 */
-			success: false;
-
-			/**
-			 * Stats about the planning process
-			 */
-			stats: PlanningStats;
-	  };
 
 export interface PlannerConfig {
 	/**
